Replaced anti special action
it has now become anti-jetpack hack due to bugs that formed.
AddPlayerClass is currently not hooked properly either due to crashes
made (which hopefully I'll fix soon), and lastly the jetpack detection
will not trigger if you use spawn a jetpack on a player via the server,
which is nice.