Update vehiclespeedlist.var
Some of the vehicles had too low speed for the magic number '170.0'. The vehicles that are commented we're edited, NOT ALL we're edited, some I just put the names so I can know which vehicles they are.

Roadtrain, sabre, flatbed, squalo, yankee, blade, savanna I remember changing

There is one last issue. Bikes. As we all know bike speeds are a bit stupid in GTA and there is something called the 'wheelie bug'. I tested PCJ-600 going at 250 kph with 170 as the magic number with this wheelie bug. It CAN'T get over 200 kph without it. This goes for all sport motor bikes. I did not edit the values of these bikes for now. Wheelies can be detected, I have done it using rotation quats in a gamemode. If we really want a different speed for wheelies ONLY. I will do it.